Move unused single topology code to attic
