+++ attic/single_topology/README.md
@@ -1,0 +1,8 @@
+# Legacy code for identifying dummy groups and anchors in single topology transformations
+
+The code here was mostly introduced in [#633](https://github.com/proteneer/timemachine/pull/633)
+and refined in [#656](https://github.com/proteneer/timemachine/pull/656).
+
+Simplified single topology, introduced in [#756](https://github.com/proteneer/timemachine/pull/756)
+utilizes only a subset of the original functionality; the remainder was moved here to simplify maintenance.
+
